Time: 48 ms (80.98%), Space: 6 MB (47.83%) - LeetHub

+++ 1680-concatenation-of-consecutive-binary-numbers/1680-concatenation-of-consecutive-binary-numbers.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int concatenatedBinary(int n) {
         
-        long long res = 0;
+        long res = 0;
         int const mod = 1e9 + 7;
         
         for(int i=1;i<=n;i++){
